# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentProvider = typeof AgentProvider[keyof typeof AgentProvider];


export const AgentProvider = {
  anthropic: 'anthropic',
  openai: 'openai',
  gemini: 'gemini',
  openrouter: 'openrouter',
} as const;

export type AgentRole = typeof AgentRole[keyof typeof AgentRole];


export const AgentRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AgentMessage {
  role: AgentRole;
  content: string;
}

export interface AgentFile {
  path: string;
  content: string;
}

export interface AgentAskBody {
  provider: AgentProvider;
  /**
     * @minItems 1
     * @maxItems 20
     */
  messages: AgentMessage[];
  /** @maxItems 80 */
  files: AgentFile[];
}

export interface ConfigureAgentKeyInput {
  provider: AgentProvider;
  /**
     * @minLength 1
     * @maxLength 500
     */
  apiKey: string;
}

export interface AgentToolEvent {
  tool: string;
  input: string;
  result: string;
}

export interface AgentStatus {
  provider: AgentProvider;
  model: string;
  configured: boolean;
  message: string;
}

export interface AgentResponse {
  provider: AgentProvider;
  model: string;
  message: string;
  events: AgentToolEvent[];
}

export type GetAgentStatusParams = {
provider: AgentProvider;
};