# lib/api-zod/src/generated/types/configureAgentKeyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentProvider } from './agentProvider';

export interface ConfigureAgentKeyInput {
  provider: AgentProvider;
  /**
     * @minLength 1
     * @maxLength 500
     */
  apiKey: string;
}